Extended sensors capabilities to be heading-based (via compass)

+++ Networking/NaiveNetworkHandler.h
@@ -6,9 +6,9 @@
 class NaiveNetworkHandler : public whytsoft::NetworkHandler
 {
 public:
-    explicit NaiveNetworkHandler(size_t socketId) : NetworkHandler{socketId} { }
-    virtual void onNewConnection(size_t socketId) const override;
-	virtual void onConnectionClosed(size_t socketId) const override;
+    explicit NaiveNetworkHandler(short unsigned int socketId) : NetworkHandler{socketId} { }
+    virtual void onNewConnection(long long unsigned int socketId) const override;
+	virtual void onConnectionClosed(long long unsigned int socketId) const override;
     ~NaiveNetworkHandler() = default;
 private:
 };
